modify : api kadai3

+++ src/jp/co/goalist/api/kadai3.java
@@ -112,7 +112,7 @@ public class Kadai3{
             SimpleDateFormat df = new SimpleDateFormat("yyyy/MM/dd");
             while(m.find()){
                 
-                String limit = m.group(10).equals("0")? "unlimited" : df.format(new Date(Integer.parseInt(m.group(10)) * 1000));
+                String limit = m.group(10).equals("0")? "unlimited" : df.format(new Date((long)Integer.parseInt(m.group(10)) * 1000));
                 
                 bw.write( "\"" + m.group(1) + "\","
                         + "\"" + convertToOiginal(m.group(3)) + "\","
